backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4

sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0

python-multipart==0.0.20

pypdf==4.3.1
python-pptx==1.0.2